backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

# Phase 4 - Export/Print/Performance packages
openpyxl==3.1.2
reportlab==4.0.7
psutil==5.9.6
